+++ framework/src/main/java/org/checkerframework/framework/source/SourceChecker.java
@@ -547,10 +547,10 @@ public abstract class SourceChecker extends AbstractTypeProcessor implements Opt
 
     // Keep in sync with check in checker-framework/build.gradle .
     int jreVersion = SystemUtil.jreVersion;
-    if (jreVersion != 8 && jreVersion != 11 && jreVersion != 17 && jreVersion != 18) {
+    if (jreVersion != 8 && jreVersion != 11 && jreVersion != 17 && jreVersion != 19) {
       message(
           (hasOption("permitUnsupportedJdkVersion") ? Kind.NOTE : Kind.WARNING),
-          "The Checker Framework is tested with JDK 8, 11, 17, and 18."
+          "The Checker Framework is tested with JDK 8, 11, 17, and 19."
               + " You are using version %d.",
           jreVersion);
     }