+++ docs/CHANGELOG.md
@@ -1,4 +1,4 @@
-Version 3.27.1 (December1, 2022)
+Version 3.28.0 (December 1, 2022)
 ---------------------------------
 
 **User-visible changes:**
@@ -6,6 +6,8 @@ Version 3.27.1 (December1, 2022)
 The Checker Framework runs under JDK 19 -- that is, it runs on a version 19 JVM.
 
 **Implementation details:**
+
+Renamed `TryFinallyScopeCell` to `LabelCell`.
 
 **Closed issues:**